cli: Add release ID to ps output
Some users have had issues where scaling does not affect running jobs,
and it is likely that the jobs are from an old release which failed to
scale down.

Adding the release ID to ps output will help to identify this issue and
make it easier to scale the given release down (using `flynn scale -r`).

Signed-off-by: Lewis Marshall <lewis@lmars.net>